Use copy assignment modal in lessons
